+++ src/balaambot/bot_commands/music_commands.py
@@ -8,8 +8,9 @@ from discord.ui import Button, View
 
 from balaambot import discord_utils
 from balaambot.utils import sec_to_string
+from balaambot.youtube import audio as yt_audio
+from balaambot.youtube import jobs as yt_jobs
 from balaambot.youtube import utils as yt_utils
-from balaambot.youtube import youtube_audio, youtube_jobs
 
 logger = logging.getLogger(__name__)
 
@@ -118,7 +119,7 @@ class MusicCommands(commands.Cog):
             return
         vc, mixer = vc_mixer
 
-        results = await youtube_audio.search_youtube(query)
+        results = await yt_audio.search_youtube(query)
         # Each result is a tuple: (url, title, duration_in_seconds)
 
         if not results:
@@ -156,7 +157,7 @@ class MusicCommands(commands.Cog):
         vc, mixer = vc_mixer
 
         # Fetch playlist video URLs
-        track_urls = await youtube_audio.get_playlist_video_urls(playlist_url)
+        track_urls = await yt_audio.get_playlist_video_urls(playlist_url)
         if not track_urls:
             return await interaction.followup.send(
                 "Failed to fetch playlist or playlist is empty.", ephemeral=True
@@ -165,7 +166,7 @@ class MusicCommands(commands.Cog):
         # Enqueue all tracks and start background fetches
         for track_url in track_urls:
             # These have to be awaited, to preserve order.
-            await youtube_jobs.add_to_queue(
+            await yt_jobs.add_to_queue(
                 vc, track_url, text_channel=interaction.channel_id
             )
 
@@ -185,9 +186,9 @@ class MusicCommands(commands.Cog):
         vc, mixer = vc_mixer
 
         # Add to queue. Playback (in mixer) will await cache when it's time
-        await youtube_jobs.add_to_queue(vc, url, text_channel=interaction.channel_id)
+        await yt_jobs.add_to_queue(vc, url, text_channel=interaction.channel_id)
 
-        track_meta = await youtube_audio.get_youtube_track_metadata(url)
+        track_meta = await yt_audio.get_youtube_track_metadata(url)
         if track_meta is None:
             await interaction.followup.send(
                 f"Failed to fetch track metadata. Please check the URL. [{url}]",
@@ -195,7 +196,7 @@ class MusicCommands(commands.Cog):
             )
             return
 
-        queue = await youtube_jobs.list_queue(vc)
+        queue = await yt_jobs.list_queue(vc)
         pos = len(queue)
 
         runtime = track_meta["runtime_str"]
@@ -231,7 +232,7 @@ class MusicCommands(commands.Cog):
             interaction.guild,
             member.voice.channel,
         )
-        upcoming = await youtube_jobs.list_queue(vc)
+        upcoming = await yt_jobs.list_queue(vc)
 
         if not upcoming:
             msg = "The queue is empty."
@@ -241,7 +242,7 @@ class MusicCommands(commands.Cog):
 
             for i, url in enumerate(upcoming[:10]):
                 new_line = f"{i + 1}. [Invalid track URL]({url})"
-                track_meta = await youtube_audio.get_youtube_track_metadata(url)
+                track_meta = await yt_audio.get_youtube_track_metadata(url)
                 if track_meta is not None:
                     new_line = (
                         f"{i + 1}. {track_meta['title']} ({track_meta['runtime_str']})"
@@ -249,7 +250,7 @@ class MusicCommands(commands.Cog):
                     total_runtime += track_meta["runtime"]
                 lines.append(new_line)
 
-            track_meta = await youtube_audio.get_youtube_track_metadata(upcoming[0])
+            track_meta = await yt_audio.get_youtube_track_metadata(upcoming[0])
             if track_meta is None:
                 lines[0] = f"**Now playing:** [Invalid track URL]({upcoming[0]})"
             else:
@@ -293,17 +294,17 @@ class MusicCommands(commands.Cog):
             interaction.guild,
             member.voice.channel,
         )
-        await youtube_jobs.skip(vc)
+        await yt_jobs.skip(vc)
         logger.info("Skipped track for guild_id=%s", interaction.guild.id)
 
-        track_url = youtube_jobs.get_current_track(vc)
+        track_url = yt_jobs.get_current_track(vc)
         if not track_url:
             await interaction.followup.send(
                 "No track is currently playing.", ephemeral=True
             )
             return
 
-        track_meta = await youtube_audio.get_youtube_track_metadata(track_url)
+        track_meta = await yt_audio.get_youtube_track_metadata(track_url)
         if track_meta is None:
             await interaction.followup.send(
                 f"Failed to fetch track metadata. Please check the URL. [{track_url}]",
@@ -341,7 +342,7 @@ class MusicCommands(commands.Cog):
             interaction.guild,
             member.voice.channel,
         )
-        await youtube_jobs.stop(vc)
+        await yt_jobs.stop(vc)
         await interaction.response.send_message(
             "⏹️    Stopped and cleared YouTube queue.", ephemeral=False
         )
@@ -371,8 +372,8 @@ class MusicCommands(commands.Cog):
         logger.info("Clearing YouTube queue for guild_id=%s", interaction.guild.id)
 
         # Clear the queue
-        await youtube_jobs.clear_queue(vc)
-        current_queue = await youtube_jobs.list_queue(vc)
+        await yt_jobs.clear_queue(vc)
+        current_queue = await yt_jobs.list_queue(vc)
         logger.info("queue after clearing: %s", current_queue)
 
         await interaction.response.send_message(
